logica del producto aplicada

+++ src/main/java/com/apirest/facturacionrest/model/Producto.java
@@ -13,38 +13,46 @@ public class Producto {
     private String nombre;
     private Double precio;
     private Double impuesto;
-    private Double stock;
-    
+    private Integer stock;
+
     public Long getId() {
         return id;
     }
+
     public void setId(Long id) {
         this.id = id;
     }
+
     public String getNombre() {
         return nombre;
     }
+
     public void setNombre(String nombre) {
         this.nombre = nombre;
     }
+
     public Double getPrecio() {
         return precio;
     }
+
     public void setPrecio(Double precio) {
         this.precio = precio;
     }
+
     public Double getImpuesto() {
         return impuesto;
     }
+
     public void setImpuesto(Double impuesto) {
         this.impuesto = impuesto;
     }
-    public Double getStock() {
+
+    public Integer getStock() {
         return stock;
     }
-    public void setStock(Double stock) {
+
+    public void setStock(Integer stock) {
         this.stock = stock;
     }
 
-    
 }
